+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ For git tags, official releases are marked like "v1.0.0", while unofficial ones 
 
 ### Fixed
 
+* Click "New Text" on the first page of ``do_feeds.php`` had no effect. Changed to "New Feed".
 * Step 2 of feed wizard was sometimes failing because ``$_SESSION['wizard']['host']`` was a string an not an array.
 
 ## 2.7.0-fork (March 14 2023)